Move helpers to test vm util
